+++ skoolkit/loadsample.py
@@ -1,0 +1,59 @@
+# Copyright 2022 Richard Dymond (rjdymond@gmail.com)
+#
+# This file is part of SkoolKit.
+#
+# SkoolKit is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# SkoolKit is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# SkoolKit. If not, see <http://www.gnu.org/licenses/>.
+
+class Accelerator:
+    def __init__(self, code, in_time, loop_time, loop_r_inc, ear_mask):
+        self.code = code
+        self.in_time = in_time
+        self.loop_time = loop_time
+        self.loop_r_inc = loop_r_inc
+        self.ear_mask = ear_mask
+
+ACCELERATORS = {
+    'microsphere': Accelerator(
+        [               # LD_SAMPLE INC B          [4]
+            0xC8,       #           RET Z          [11/5]
+            0x3E, 0x7F, #           LD A,$7F       [7]
+            0xDB, 0xFE, #           IN A,($FE)     [11]
+            0x1F,       #           RRA            [4]
+            0xA7,       #           AND A          [4]
+            0xA9,       #           XOR C          [4]
+            0xE6, 0x20, #           AND $20        [7]
+            0x28, 0xF3  #           JR Z,LD_SAMPLE [12/7]
+        ],
+        16,   # 16 T-states until first IN A,($FE)
+        58,   # 58 T-states per loop iteration
+        9,    # R register increment per loop iteration
+        0x20  # EAR mask
+    ),
+
+    'rom': Accelerator(
+        [               # LD_SAMPLE INC B          [4]
+            0xC8,       #           RET Z          [11/5]
+            0x3E, 0x7F, #           LD A,$7F       [7]
+            0xDB, 0xFE, #           IN A,($FE)     [11]
+            0x1F,       #           RRA            [4]
+            0xD0,       #           RET NC         [11/5]
+            0xA9,       #           XOR C          [4]
+            0xE6, 0x20, #           AND $20        [7]
+            0x28, 0xF3  #           JR Z,LD_SAMPLE [12/7]
+        ],
+        16,   # 16 T-states until first IN A,($FE)
+        59,   # 59 T-states per loop iteration
+        9,    # R register increment per loop iteration
+        0x20  # EAR mask
+    ),
+}